requirements.txt
# FILEPATH: /home/ubuntu/RRAP/requirements.txt
praw=7.7.1
python==3.10.12
python-dotenv==1.0.0

